+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[1.4.0]: https://github.com/johnarp/warbond-tracker/compare/v1.3.5...v1.4.0
 [1.3.5]: https://github.com/johnarp/warbond-tracker/compare/v1.3.4...v1.3.5
 [1.3.4]: https://github.com/johnarp/warbond-tracker/compare/v1.3.3...v1.3.4
 [1.3.3]: https://github.com/johnarp/warbond-tracker/compare/v1.3.2...v1.3.3
@@ -34,6 +35,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.1.2]: https://github.com/johnarp/warbond-tracker/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/johnarp/warbond-tracker/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/johnarp/warbond-tracker/releases/tag/0.1.0
+
+## [1.4.0] - 2026-04-30
+
+### Added
+
+- Medals spent tracker for each Warbond
+- Medals text field in pop-up modal
+- Show Medals toggle
+- Sort by fewest/most medals left options
+- Total Medals Spent section in Profile
+
+### Changed
+
+- Sort, Type, and Status options are saved and can be imported/exported
+- README preview image
 
 ## [1.3.5] - 2026-04-29